refactor: Model 폴더 생성
