+++ src/Model/BridgeStep.js
